requirements.txt
gcloud

python3
pyyaml

R
